FinPro/requirements.txt
pandas>=1.5.0
numpy>=1.23.0
pdfplumber>=0.9.0
spacy>=3.5.0
matplotlib>=3.7.0
seaborn>=0.12.0
regex>=2023.5.5
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0.tar.gz
streamlit
plotly
pandas
spacy
pdfplumber
pytesseract
opencv-python
Pillow
